requirements.txt
openai
chromadb
sentence-transformers
torch
numpy<2.0
psycopg2-binary
sqlparse
pydantic>=2.0
pydantic-settings